--- a/crates/nu-cli/src/completions/variable_completions.rs
+++ b/crates/nu-cli/src/completions/variable_completions.rs
@@ -262,9 +262,9 @@ fn nested_suggestions(
 
             output
         }
-        Value::LazyRecord { val, span } => {
+        Value::LazyRecord { val, .. } => {
             // Add all the columns as completion
-            for (column_name, _) in val.get_column_map() {
+            for column_name in val.columns() {
                 output.push(Suggestion {
                     value: column_name,
                     description: None,