--- a/crates/nu-command/src/filters/columns.rs
+++ b/crates/nu-command/src/filters/columns.rs
@@ -122,6 +122,15 @@ fn getcol(
                 .into_pipeline_data(ctrlc)
                 .set_metadata(metadata))
         }
+        PipelineData::Value(Value::LazyRecord { val, span }, ..) => Ok(val
+            .columns()
+            .into_iter()
+            .map(move |x| Value::String {
+                val: x.into(),
+                span: head,
+            })
+            .into_pipeline_data(ctrlc)
+            .set_metadata(metadata)),
         PipelineData::Value(Value::Record { cols, .. }, ..) => Ok(cols
             .into_iter()
             .map(move |x| Value::String { val: x, span: head })