--- a/crates/nu-command/src/formats/to/json.rs
+++ b/crates/nu-command/src/formats/to/json.rs
@@ -137,6 +137,7 @@ pub fn value_to_json_value(v: &Value) -> Result<nu_json::Value, ShellError> {
             nu_json::Value::Object(m)
         }
         Value::CustomValue { val, .. } => val.to_json(),
+        Value::LazyRecord { val, span } => todo!(),
     })
 }