--- a/crates/nu-command/src/formats/to/json.rs
+++ b/crates/nu-command/src/formats/to/json.rs
@@ -136,8 +136,11 @@ pub fn value_to_json_value(v: &Value) -> Result<nu_json::Value, ShellError> {
             }
             nu_json::Value::Object(m)
         }
+        Value::LazyRecord { val, .. } => {
+            let collected = val.collect()?;
+            value_to_json_value(&collected)?
+        },
         Value::CustomValue { val, .. } => val.to_json(),
-        Value::LazyRecord { val, span } => todo!(),
     })
 }