--- a/crates/nu-command/src/formats/to/nuon.rs
+++ b/crates/nu-command/src/formats/to/nuon.rs
@@ -188,6 +188,7 @@ fn value_to_string(v: &Value, span: Span) -> Result<String, ShellError> {
         // All strings outside data structures are quoted because they are in 'command position'
         // (could be mistaken for commands by the Nu parser)
         Value::String { val, .. } => Ok(escape_quote_string(val)),
+        Value::LazyRecord { val, span } => todo!(),
     }
 }