--- a/crates/nu-command/src/formats/to/nuon.rs
+++ b/crates/nu-command/src/formats/to/nuon.rs
@@ -188,7 +188,7 @@ pub fn value_to_string(v: &Value, span: Span) -> Result<String, ShellError> {
         Value::LazyRecord { val, .. } => {
             let collected = val.collect()?;
             value_to_string(&collected, span)
-        },
+        }
         // All strings outside data structures are quoted because they are in 'command position'
         // (could be mistaken for commands by the Nu parser)
         Value::String { val, .. } => Ok(escape_quote_string(val)),