--- a/crates/nu-command/src/formats/to/text.rs
+++ b/crates/nu-command/src/formats/to/text.rs
@@ -147,6 +147,7 @@ fn local_into_string(value: Value, separator: &str, config: &Config) -> String {
         Value::Binary { val, .. } => format!("{:?}", val),
         Value::CellPath { val, .. } => val.into_string(),
         Value::CustomValue { val, .. } => val.value_string(),
+        Value::LazyRecord { val, span } => todo!(),
     }
 }