--- a/crates/nu-command/src/formats/to/text.rs
+++ b/crates/nu-command/src/formats/to/text.rs
@@ -140,11 +140,9 @@ fn local_into_string(value: Value, separator: &str, config: &Config) -> String {
             .map(|(x, y)| format!("{}: {}", x, local_into_string(y.clone(), ", ", config)))
             .collect::<Vec<_>>()
             .join(separator),
-        Value::LazyRecord { val, .. } => {
-            match val.collect() {
-                Ok(val) => local_into_string(val, separator, config),
-                Err(error) => format!("{:?}", error),
-            }
+        Value::LazyRecord { val, .. } => match val.collect() {
+            Ok(val) => local_into_string(val, separator, config),
+            Err(error) => format!("{:?}", error),
         },
         Value::Block { val, .. } => format!("<Block {}>", val),
         Value::Closure { val, .. } => format!("<Closure {}>", val),