--- a/crates/nu-engine/src/lib.rs
+++ b/crates/nu-engine/src/lib.rs
@@ -4,6 +4,7 @@ pub mod documentation;
 pub mod env;
 mod eval;
 mod glob_from;
+mod nu_variable;
 mod scope;
 
 pub use call_ext::CallExt;