--- a/crates/nu-engine/src/nu_variable.rs
+++ b/crates/nu-engine/src/nu_variable.rs
@@ -7,6 +7,7 @@ use serde::{Deserialize, Serialize};
 // a CustomValue for the special $nu variable
 // $nu used to be a plain old Record, but CustomValue lets us load different fields/columns lazily. This is important for performance;
 // collecting all the information in $nu is expensive and unnecessary if  you just want a subset of the data
+// should this use #[typetag::serde] instead?
 #[derive(Serialize, Deserialize)]
 pub struct NuVariable {
     #[serde(skip)]