… Cargo.lock
… crates/nu-command/tests/format_conversions/nuon.rs
… crates/nu-engine/Cargo.toml
… crates/nu-plugin/src/plugin/declaration.rs
… crates/nu-protocol/src/shell_error.rs
… crates/nu-protocol/src/value/lazy_record.rs
… crates/nu-protocol/src/value/mod.rs