@fdncred 2022-12-28
It's interesting that running `$nu` consecutively displays that results in different orders.
<img width="979" alt="Screenshot 2022-12-27 at 6 49 40 PM" src="https://user-images.githubusercontent.com/343840/209740901-207a2c7a-e8a4-4143-a945-cedfa2278fae.png">