@rgwood 2022-12-28
Ah, good catch - the order of Rust hashmaps is (understandably) not well-defined.

> Not trying to be critical but it seems a tiny bit slower when doing $nu by a few ms

How are you measuring that? `benchmark`?