@fdncred 2022-12-28
> How are you measuring that? benchmark? 

Like this. In my prompt on the right side (to the left of the time) I have the `$env.CMD_DURATION_MS`. Just using that.
<img width="852" alt="Screenshot 2022-12-27 at 7 11 45 PM" src="https://user-images.githubusercontent.com/343840/209742256-4065a6d7-e349-40b1-9369-94c49e74c282.png">