@rgwood 2022-12-28
Gotcha. I'm seeing `$nu` take roughly 3-4ms on `main` and 4-5ms on this PR. The clone of `engine_state` and `stack` might be responsible for a bit of the slowdown relative to the default.