@fdncred 2022-12-28
> 3-4ms on main and 4-5ms on this PR

On my mac, main is 9-10ms and 11-12ms on this PR. 

I'm wondering why mine is 3x yours. I'm just guessing that I have more stuff that I'm sourcing and you're just using defaults?