@rgwood 2022-12-28
> I'm wondering why mine is 3x yours. I'm just guessing that I have more stuff that I'm sourcing and you're just using defaults?

Probably that and the fact that I'm running a desktop CPU that consumes more electricity than a small town.

> I'm also not convinced the PR is loading everything, e.g. 516 commands on main vs 397 commands on this PR.

Weird! I don't see any difference, let me know if you can figure out a pattern.