@fdncred 2022-12-28
> Weird! I don't see any difference, let me know if you can figure out a pattern.

Ugh. I found it and it's my mistake. I built your PR *without* dataframes and my regular nu is *with* dataframes. At least it's not your code 🤣 


Now my perf is ~14ms without the PR and ~13ms with this PR - so it's faster-ish. So sorry for all the misleading perf metrics.

So, I think you can ignore all my comments except the out of order results.