@rgwood 2022-12-29
I'll try redoing this with a `vec` of `OnceCell<Result<Value, ShellError>>` items; that should address the ordering issue and avoid evaluating columns twice.