@kubouch 2022-12-31
Very cool! I'd like to see that working.

My main concern is storing the engine state and stack inside the value. Somehow I feel like it goes a bit against Nushell's design where we always tag these objects along wherever we're doing something with Values so that the Values do not need to store any sort of state inside of them.

For the HashMap order problem, see IndexMap (used in https://github.com/nushell/nushell/blob/main/crates/nu-protocol/src/module.rs). For some reason JT wasn't keen on using that but it might be the easiest way to solve the ordering problem.