@rgwood 2023-01-02
Spent a bit more time on this today. `IndexMap` seems like a good way to solve the ordering problem.

I haven't found a great way to avoid storing engine state and stack inside the value. I think Jakub's suggestion is the most promising:

> An alternative: Don't store the EngineState and Stack at all and instead have `get_column_map(&self, engine_state: &EngineState, stack: &Stack)`. Could we do that?

That's probably doable but it involves threading EngineState and Stack through a lot of cell path code that is actively being worked on for error handling. I might wait until the cell path stuff settles down before picking this up again.