@rgwood 2023-01-06
Passing EngineState and Stack everywhere that `LazyRecord` might be evaluated is unfortunately not very practical. Cell path access sometimes happens in code without access to EngineState and Stack (I should have written down exactly where, but I forgot).

I have also been trying to rewrite this to use `OnceCell`; it would be nice if column values were only evaluated once. That's probably possible, but it gets very tricky with lifetimes (the cell initializer functions need to keep a reference to NuValue) and it requires more fighting with rustc than I have the energy for right now.

I'm going to drop this for now. In the short term, if we want to speed up `$nu` I would suggest just moving the `scope` information elsewhere.