@rgwood 2023-01-11
OK, I can't stay away from this. I've simplified the `LazyRecord` trait and cleaned things up a bit. If we can get tentative agreement on the way forward, I'll try to polish this up and get it fully ready for code review. One point that I would like to seek consent on:

### I think the `$nu` implementation of LazyRecord needs to take ownership of EngineState and Stack.

IMO this is important for ease of implementation _and_ correctness. Consider a situation where a user assigns `$nu` to a variable and then accesses it later:

```bash
let foo = $nu
alias some_alias = some_command
let bar = $foo.scope
```

Before LazyRecord, `bar` would not contain `some_alias`. But if LazyRecord references the "living" EngineState, `bar` _will_ contain `some_alias`.

I don't think we should change the existing behaviour; as much as possible, laziness should be an implementation detail that is unobservable to users. Taking ownership of `EngineState` solves this nicely by using a snapshot of EngineState from the time the LazyRecord was created.