@kubouch 2023-01-11
You're totally, right, we need to access the engine state / stack from the point when the variable was created. I can't think of a different way to do it it than storing a copy.