@rgwood 2023-01-19
Alright, let's give it a try.